+++ packages/slate/src/utils/deep-equal.ts
@@ -24,7 +24,6 @@ export const isDeepEqual = (
       for (let i = 0; i < a.length; i++) {
         if (a[i] !== b[i]) return false
       }
-      return true
     } else if (a !== b) {
       return false
     }